+++ lib/ui/appointments/AppointmentCard.dart
@@ -57,105 +57,152 @@ class _AppointmentCardState extends State<AppointmentCard> implements Notificati
   @override
   Widget build(BuildContext context) {
     const double imageSize = 64;
+    const double iconSize = 18;
     String? imageKey = widget.appointment.imageKey;
-    bool isFavorite = Auth2().isFavorite(widget.appointment);
-    bool starVisible = Auth2().canFavorite && widget.appointment.isUpcoming;
+    String semanticsImageLabel = 'appointment image';
+    String semanticsImageHint = 'Double tap to expand image';
 
-    return InkWell(
-        onTap: _onTapAppointmentCard,
-        child: ClipRRect(
-            borderRadius: BorderRadius.vertical(bottom: Radius.circular(4)),
-            child: Stack(children: [
-              Container(
-                  decoration: BoxDecoration(
-                      color: Styles().colors!.surface,
-                      border: Border.all(color: Styles().colors!.surfaceAccent!, width: 1),
-                      borderRadius: BorderRadius.all(Radius.circular(4))),
-                  child: Padding(
-                      padding: EdgeInsets.all(16),
-                      child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-                        Row(children: [
-                          Expanded(
-                              child: Text(StringUtils.ensureNotEmpty(widget.appointment.category),
-                                  style:Styles().textStyles?.getTextStyle("widget.item.small.semi_fat"))),
-                          Visibility(
-                              visible: starVisible,
-                              child: Semantics(
-                                  container: true,
-                                  child: Container(
-                                      child: Semantics(
-                                          label: isFavorite
-                                              ? Localization().getStringEx('widget.card.button.favorite.off.title', 'Remove From Favorites')
-                                              : Localization().getStringEx('widget.card.button.favorite.on.title', 'Add To Favorites'),
-                                          hint: isFavorite
-                                              ? Localization().getStringEx('widget.card.button.favorite.off.hint', '')
-                                              : Localization().getStringEx('widget.card.button.favorite.on.hint', ''),
-                                          button: true,
-                                          child: GestureDetector(
-                                              behavior: HitTestBehavior.opaque,
-                                              onTap: _onTapExploreCardStar,
-                                              child: Container(
-                                                  child: Padding(
-                                                      padding: EdgeInsets.only(left: 24, bottom: 5),
-                                                      child: Styles().images?.getImage(isFavorite ? 'star-filled' : 'star-outline-gray', excludeFromSemantics: true))))))))
-                        ]),
-                        Padding(
-                            padding: EdgeInsets.only(top: 6),
-                            child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-                              Row(crossAxisAlignment: CrossAxisAlignment.start, children: [
-                                Expanded(
-                                    child: Text(StringUtils.ensureNotEmpty(widget.appointment.title),
-                                        style: Styles().textStyles?.getTextStyle("widget.title.large.extra_fat"))),
-                                Visibility(
-                                    visible: StringUtils.isNotEmpty(imageKey),
-                                    child: Padding(
-                                        padding: EdgeInsets.only(left: 16, bottom: 4),
-                                        child: Semantics(
-                                            label: "appointment image",
-                                            button: true,
-                                            hint: "Double tap to expand image",
-                                            child: SizedBox(
-                                                width: imageSize,
-                                                height: imageSize,
-                                                child: InkWell(
-                                                    onTap: () => _onTapCardImage(imageKey),
-                                                    child: Styles().images?.getImage(imageKey,
-                                                        excludeFromSemantics: true,
-                                                        fit: BoxFit.fill,
-                                                        networkHeaders: Config().networkAuthHeaders))))))
-                              ]),
-                              Padding(
-                                  padding: EdgeInsets.only(top: 5),
-                                  child: Row(children: [
-                                    Padding(padding: EdgeInsets.only(right: 6), child: Styles().images?.getImage('calendar', excludeFromSemantics: true)),
-                                    Expanded(
-                                        child: Text(StringUtils.ensureNotEmpty(widget.appointment.displayScheduleTime),
-                                            style: Styles().textStyles?.getTextStyle("widget.item.regular")))
-                                  ])),
-                              Padding(
-                                  padding: EdgeInsets.only(top: 5),
-                                  child: Row(children: [
-                                    Padding(
-                                        padding: EdgeInsets.only(right: 6),
-                                        child: Styles().images?.getImage((widget.appointment.type == AppointmentType.online) ?
-                                          'laptop' : 'location', excludeFromSemantics: true)),
-                                    Expanded(
-                                        child: Text(widget.appointment.displayType ?? '',
-                                            overflow: TextOverflow.ellipsis,
-                                            style: Styles().textStyles?.getTextStyle("widget.item.regular"))),
-                                    Visibility(
-                                        visible: (widget.appointment.cancelled == true),
-                                        child: Padding(
-                                            padding: EdgeInsets.only(left: 7),
-                                            child: Text(Localization().getStringEx('widget.appointment.card.cancelled.label', 'Cancelled'),
-                                                style: Styles().textStyles?.getTextStyle("panel.appointment_detail.title.large"))))
-                                  ]))
-                            ]))
-                      ]))),
-              Container(
-                  color: (widget.appointment.isUpcoming ? Styles().colors?.fillColorSecondary : Styles().colors?.fillColorPrimary),
-                  height: 4)
-            ])));
+    String typeImageKey = (widget.appointment.type == AppointmentType.online) ? 'laptop' : 'location';
+    Color? headerColor = (widget.appointment.isUpcoming ? Styles().colors?.fillColorSecondary : Styles().colors?.fillColorPrimary);
+    String? displayTime = widget.appointment.displayShortScheduleTime;
+    String? displayType = widget.appointment.displayType;
+    String? displayHost = widget.appointment.host?.displayName;
+
+    bool isFavorite = Auth2().isFavorite(widget.appointment);
+    bool canFavorite = Auth2().canFavorite && widget.appointment.isUpcoming;
+    String favImageKey = isFavorite ? 'star-filled' : 'star-outline-gray';
+    String semanticsFavLabel = isFavorite ?
+      Localization().getStringEx('widget.card.button.favorite.off.title', 'Remove From Favorites') :
+      Localization().getStringEx('widget.card.button.favorite.on.title', 'Add To Favorites');
+    String semanticsFavHint = isFavorite ?
+      Localization().getStringEx('widget.card.button.favorite.off.hint', '') :
+      Localization().getStringEx('widget.card.button.favorite.on.hint', '');
+
+    return InkWell(onTap: _onTapAppointmentCard,child:
+      ClipRRect(borderRadius: BorderRadius.vertical(bottom: Radius.circular(4)), child:
+        Stack(children: [
+          Container(
+            decoration: BoxDecoration(
+              color: Styles().colors!.surface,
+              border: Border.all(color: Styles().colors!.surfaceAccent!, width: 1),
+              borderRadius: BorderRadius.all(Radius.circular(4))
+            ),
+            child: Padding(padding: EdgeInsets.only(left: 16, bottom: 16), child:
+              Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
+                Row(children: [
+                  Expanded(child:
+                    Padding(padding: EdgeInsets.only(top: 16, bottom: 16, right: canFavorite ? 0 : 16), child:
+                      Text(widget.appointment.category ?? '', style:
+                        Styles().textStyles?.getTextStyle("widget.item.small.semi_fat")
+                      ),
+                    ),
+                  ),
+                  Visibility(visible: canFavorite, child:
+                    Semantics(label: semanticsFavLabel, hint: semanticsFavHint, button: true, child:
+                      GestureDetector(behavior: HitTestBehavior.opaque, onTap: _onTapExploreCardStar, child:
+                        Padding(padding: EdgeInsets.all(16), child:
+                          Styles().images?.getImage(favImageKey, excludeFromSemantics: true)
+                        )
+                      )
+                    )
+                  )
+                ]),
+                
+                Padding(padding: EdgeInsets.only(right: 16), child:
+                  Row(crossAxisAlignment: CrossAxisAlignment.start, children: [
+                    Expanded(child:
+                      Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
+
+                        Text(widget.appointment.title ?? '', style:
+                          Styles().textStyles?.getTextStyle("widget.title.large.extra_fat")
+                        ),
+
+                        Visibility(visible: StringUtils.isNotEmpty(displayTime), child:
+                          Padding(padding: EdgeInsets.only(top: 4), child:
+                            Row(crossAxisAlignment: CrossAxisAlignment.start, children: [
+                              Padding(padding: EdgeInsets.only(right: 6), child: 
+                                SizedBox(width: iconSize, height: iconSize, child:
+                                  Styles().images?.getImage('calendar', excludeFromSemantics: true)
+                                )
+                              ),
+                              Expanded(child:
+                                Text(displayTime ?? '', style:
+                                  Styles().textStyles?.getTextStyle("widget.item.regular")
+                                )
+                              )
+                            ])
+                          ),
+                        ),
+                        
+                        Visibility(visible: StringUtils.isNotEmpty(displayHost), child:
+                          Padding(padding: EdgeInsets.only(top: 4), child:
+                            Row(crossAxisAlignment: CrossAxisAlignment.start, children: [
+                              Padding(padding: EdgeInsets.only(right: 6), child:
+                                SizedBox(width: iconSize, height: iconSize, child:
+                                  Styles().images?.getImage('person', excludeFromSemantics: true)
+                                )
+                              ),
+                              Expanded(child:
+                                Text(displayHost ?? '', overflow: TextOverflow.ellipsis, style:
+                                  Styles().textStyles?.getTextStyle("widget.item.regular")
+                                )
+                              ),
+                            ])
+                          ),
+                        ),
+
+                        Visibility(visible: StringUtils.isNotEmpty(displayType), child:
+                          Padding(padding: EdgeInsets.only(top: 4), child:
+                            Row(crossAxisAlignment: CrossAxisAlignment.start, children: [
+                              Padding(padding: EdgeInsets.only(right: 6), child:
+                                SizedBox(width: iconSize, height: iconSize, child:
+                                  Styles().images?.getImage(typeImageKey, excludeFromSemantics: true)
+                                )
+                              ),
+                              Expanded(child:
+                                Text(displayType ?? '', overflow: TextOverflow.ellipsis, style:
+                                  Styles().textStyles?.getTextStyle("widget.item.regular")
+                                )
+                              ),
+                            ])
+                          ),
+                        ),
+
+                      ]),
+                    ),
+
+                    Column(crossAxisAlignment: CrossAxisAlignment.end, children: [
+                      Opacity(opacity: StringUtils.isNotEmpty(imageKey) ? 1 : 0, child:
+                        Padding(padding: EdgeInsets.only(left: 16), child:
+                          Semantics(label: semanticsImageLabel, button: true, hint: semanticsImageHint, child:
+                            InkWell(onTap: () => StringUtils.isNotEmpty(imageKey) ? _onTapCardImage(imageKey) : null, child:
+                              SizedBox(width: imageSize, height: imageSize, child:
+                                Styles().images?.getImage(imageKey, excludeFromSemantics: true, fit: BoxFit.fill, networkHeaders: Config().networkAuthHeaders)
+                              )
+                            )
+                          )
+                        )
+                      ),
+
+                      Visibility(visible: (widget.appointment.cancelled == true), child:
+                        Padding(padding: EdgeInsets.only(top: 6), child:
+                          Text(Localization().getStringEx('widget.appointment.card.cancelled.label', 'Cancelled'), textAlign: TextAlign.right, style:
+                            Styles().textStyles?.getTextStyle("panel.appointment_detail.title.large")
+                          )
+                        ),
+                      )
+
+                    ],)
+                  ]),
+                ),
+
+              ])
+            )
+          ),
+          Container(color: headerColor, height: 4)
+        ])
+      )
+    );
   }
 
   void _onTapAppointmentCard() {
@@ -165,12 +212,9 @@ class _AppointmentCardState extends State<AppointmentCard> implements Notificati
 
   void _onTapCardImage(String? imageKey) {
     Analytics().logSelect(target: 'Appointment Image');
-    Navigator.push(
-        context,
-        PageRouteBuilder(
-            opaque: false,
-            pageBuilder: (context, _, __) =>
-                ModalImagePanel(imageKey: imageKey, onCloseAnalytics: () => Analytics().logSelect(target: 'Close Image'))));
+    Navigator.push(context, PageRouteBuilder(opaque: false, pageBuilder: (context, _, __) =>
+      ModalImagePanel(imageKey: imageKey, onCloseAnalytics: () => Analytics().logSelect(target: 'Close Image')))
+    );
   }
 
   void _onTapExploreCardStar() {